Changed the activities into fragments
