+++ app/src/main/java/com/svs/hztb/Activities/HomeScreenActivity.java
@@ -1,19 +1,32 @@
 package com.svs.hztb.Activities;
 
+import android.app.FragmentManager;
+import android.content.Context;
+import android.support.v4.content.res.ResourcesCompat;
 import android.support.v4.view.GravityCompat;
 import android.support.v4.widget.DrawerLayout;
-import android.support.v7.app.AppCompatActivity;
+import android.support.v7.app.ActionBar;
 import android.os.Bundle;
 import android.view.Gravity;
+import android.view.LayoutInflater;
 import android.view.Menu;
-import android.view.MenuItem;
 import android.view.View;
+import android.view.ViewGroup;
+import android.widget.AdapterView;
+import android.widget.ImageView;
 import android.widget.ListView;
+import android.widget.RelativeLayout;
+import android.widget.TextView;
 
+import com.svs.hztb.Adapters.SlideMenuAdapter;
+import com.svs.hztb.Interfaces.IDrawerClosed;
 import com.svs.hztb.R;
 
-public class HomeScreenActivity extends AbstractActivity {
-
+public class HomeScreenActivity extends AbstractActivity implements IDrawerClosed {
+    protected SlideMenuAdapter slideMenuAdapter;
+    protected DrawerLayout mDrawerLayout;
+    protected String[] menuItems;
+    protected ListView mDrawerList;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -22,6 +35,115 @@ public class HomeScreenActivity extends AbstractActivity {
         actionBarSettingswithNavigation(R.string.title_activity_home_screen);
         intalizeDrawer();
     }
+
+
+    protected void intalizeDrawer(){
+        moveDrawerToTop();
+        mDrawerLayout = (DrawerLayout) findViewById(R.id.drawer_layout);
+        menuItems = getResources().getStringArray(R.array.side_menu_items);
+        mDrawerList = (ListView) findViewById(R.id.left_drawer);
+        // set a custom shadow that overlays the main content when the drawer opens
+        mDrawerLayout.setDrawerShadow(R.drawable.drawer_shadow, GravityCompat.START);
+        slideMenuAdapter = new SlideMenuAdapter(getApplicationContext(),menuItems,this);
+        mDrawerList.setAdapter(slideMenuAdapter);
+        mDrawerList.setOnItemClickListener(new AdapterView.OnItemClickListener() {
+            @Override
+            public void onItemClick(AdapterView<?> adapterView, View view, int i, long l) {
+                if (i == 5){
+                    if (mDrawerLayout.isDrawerOpen(Gravity.LEFT))
+                    {
+                        selectItem(5);
+                        mDrawerLayout.closeDrawer(Gravity.LEFT);
+                    }
+                }
+            }
+        });
+    }
+
+    /** Swaps fragments in the main content view */
+    private void selectItem(int position) {
+        // Create a new fragment and specify the planet to show based on position
+
+
+        NewRequestActivity fragment = new NewRequestActivity();
+        Bundle args = new Bundle();
+        fragment.setArguments(args);
+
+        // Insert the fragment by replacing any existing fragment
+        FragmentManager fragmentManager = getFragmentManager();
+        fragmentManager.beginTransaction()
+                .replace(R.id.fragment, fragment).addToBackStack("1")
+                .commit();
+
+        // Highlight the selected item, update the title, and close the drawer
+        mDrawerList.setItemChecked(position, true);
+        mDrawerLayout.closeDrawer(mDrawerList);
+    }
+
+
+
+
+    public void onDrawerClosedClicked(){
+        mDrawerLayout.closeDrawers();
+    }
+
+    private void moveDrawerToTop() {
+        LayoutInflater inflater = (LayoutInflater) getSystemService(Context.LAYOUT_INFLATER_SERVICE);
+        DrawerLayout drawer = (DrawerLayout) inflater.inflate(R.layout.decor, null); // "null" is important.
+
+        // HACK: "steal" the first child of decor view
+        ViewGroup decor = (ViewGroup) getWindow().getDecorView();
+        View child = decor.getChildAt(0);
+        decor.removeView(child);
+        RelativeLayout container = (RelativeLayout) drawer.findViewById(R.id.content_frame); // This is the container we defined just now.
+        container.addView(child, 0);
+        drawer.findViewById(R.id.left_drawer).setPadding(0, getStatusBarHeight(), 0, 0);
+        // Make the drawer replace the first child
+        decor.addView(drawer);
+    }
+    /**
+     * Action bar settings are updated
+     */
+    protected void actionBarSettingswithNavigation(int title) {
+        getSupportActionBar().setDisplayOptions(ActionBar.DISPLAY_SHOW_CUSTOM);
+        LayoutInflater mInflater = LayoutInflater.from(this);
+        View mCustomView = mInflater.inflate(R.layout.actionbar_title, null);
+        getSupportActionBar().setCustomView(mCustomView);
+
+        TextView titleView = (TextView) mCustomView.findViewById(R.id.textview_actionbarTitle);
+        titleView.setVisibility(View.INVISIBLE);
+
+        RelativeLayout actionBarWithoutNavigation =  (RelativeLayout) mCustomView.findViewById(R.id.layout_back_actionbar);
+        actionBarWithoutNavigation.setVisibility(View.VISIBLE);
+
+        final ImageView navDrawer = (ImageView)mCustomView.findViewById(R.id.button_nav);
+        navDrawer.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                mDrawerLayout.openDrawer(Gravity.LEFT);
+            }
+        });
+
+        TextView customTitle = (TextView) mCustomView.findViewById(R.id.titleView);
+        customTitle.setText(getResources().getString(title));
+
+        getSupportActionBar().setBackgroundDrawable(ResourcesCompat.getDrawable(getResources(), R.drawable.action_bar_drawble, null));
+        getSupportActionBar().setDisplayShowCustomEnabled(true);
+        getSupportActionBar().setDisplayShowTitleEnabled(false);
+    }
+
+
+
+
+    public int getStatusBarHeight() {
+        int result = 0;
+        int resourceId = getResources().getIdentifier("status_bar_height", "dimen", "android");
+        if (resourceId > 0) {
+            result = getResources().getDimensionPixelSize(resourceId);
+        }
+        return result;
+    }
+
 
     @Override
     protected void onResume() {
@@ -63,8 +185,4 @@ public class HomeScreenActivity extends AbstractActivity {
         return true;
     }
 
-    @Override
-    public void onBackPressed() {
-        alertDialog();
-    }
 }
