+++ app/src/main/java/com/svs/hztb/Adapters/SlideMenuAdapter.java
@@ -33,7 +33,6 @@ public class SlideMenuAdapter extends BaseAdapter {
         drawerClosed = iDrawerClosed;
     }
 
-
     @Override
     public int getCount() {
         return menuItems.length;
