+++ app/src/main/java/com/svs/hztb/Fragments/HomeLaunchingFragment.java
@@ -1,0 +1,21 @@
+package com.svs.hztb.Fragments;
+
+import android.os.Bundle;
+import android.support.v4.app.Fragment;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import com.svs.hztb.R;
+
+
+public class HomeLaunchingFragment extends Fragment {
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_home_launching, container, false);
+    }
+
+}
